requirements.txt
python-dotenv
PyYAML